Move shell scripts to os dir
