+++ nest-bot/src/events/app_home_opened.ts
@@ -6,7 +6,7 @@ import approved_home from "../blocks/approved_home.js";
 import unapproved_home from "../blocks/unapproved_home.js";
 import unregistered_home from "../blocks/unregistered_home.js";
 
-import get_user_shell from "../util/get_user_shell.js";
+import get_user_shell from "../os/get_user_shell.js";
 
 export async function app_home_opened(app: Slack.App) {
   app.event("app_home_opened", async ({ event, client }) => {
