+++ nest-bot/src/os/get_shells.ts
