+++ nest-bot/src/os/get_user_shell.ts
